pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bluetooth-screen-lock"
version = "0.1.17"
description = "Lock/unlock the screen based on Bluetooth proximity (GNOME/GTK tray)"
readme = "README.md"
authors = [
  { name = "Bluetooth Screen Lock Maintainers" }
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
# Keep runtime deps minimal; distro packaging (Debian) will map these to system packages
# PyGObject (gi) is provided by the OS and intentionally not listed here.
dependencies = [
  "bleak>=0.22.3",
  "PyYAML>=6.0.2",
]
keywords = ["bluetooth", "screen", "lock", "gnome", "gtk", "tray"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX :: Linux",
  "Environment :: X11 Applications :: GTK",
]

[project.scripts]
bluetooth-screen-lock = "bluetooth_screen_lock.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["bluetooth_screen_lock*"]

[tool.setuptools.package-data]
# No package data by default; icons/desktop are installed by the distro packaging.
bluetooth_screen_lock = []


